Create CellLayerBase
