+++ OpenRA.Game/Map/CellLayer.cs
@@ -10,42 +10,28 @@
 #endregion
 
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using OpenRA.Primitives;
 
 namespace OpenRA
 {
 	// Represents a layer of "something" that covers the map
-	public class CellLayer<T> : IEnumerable<T>
+	public sealed class CellLayer<T> : CellLayerBase<T>
 	{
-		public readonly Size Size;
-		readonly Rectangle bounds;
-		public readonly MapGridType GridType;
 		public event Action<CPos> CellEntryChanged = null;
 
-		readonly T[] entries;
-
 		public CellLayer(Map map)
-			: this(map.Grid.Type, new Size(map.MapSize.X, map.MapSize.Y)) { }
+			: base(map) { }
 
 		public CellLayer(MapGridType gridType, Size size)
-		{
-			Size = size;
-			bounds = new Rectangle(0, 0, Size.Width, Size.Height);
-			GridType = gridType;
-			entries = new T[size.Width * size.Height];
-		}
+			: base(gridType, size) { }
 
-		public void CopyValuesFrom(CellLayer<T> anotherLayer)
+		public override void CopyValuesFrom(CellLayerBase<T> anotherLayer)
 		{
-			if (Size != anotherLayer.Size || GridType != anotherLayer.GridType)
-				throw new ArgumentException(
-					"layers must have a matching size and shape (grid type).", "anotherLayer");
 			if (CellEntryChanged != null)
 				throw new InvalidOperationException(
 					"Cannot copy values when there are listeners attached to the CellEntryChanged event.");
-			Array.Copy(anotherLayer.entries, entries, entries.Length);
+
+			base.CopyValuesFrom(anotherLayer);
 		}
 
 		public static CellLayer<T> CreateInstance(Func<MPos, T> initialCellValueFactory, Size size, MapGridType mapGridType)
@@ -107,23 +93,6 @@ namespace OpenRA
 				if (CellEntryChanged != null)
 					CellEntryChanged(uv.ToCPos(GridType));
 			}
-		}
-
-		/// <summary>Clears the layer contents with a known value</summary>
-		public void Clear(T clearValue)
-		{
-			for (var i = 0; i < entries.Length; i++)
-				entries[i] = clearValue;
-		}
-
-		public IEnumerator<T> GetEnumerator()
-		{
-			return ((IEnumerable<T>)entries).GetEnumerator();
-		}
-
-		IEnumerator IEnumerable.GetEnumerator()
-		{
-			return GetEnumerator();
 		}
 
 		public bool Contains(CPos cell)
